requirements.txt
streamlit
ultralytics
opencv-python
Pillow
numpy
pandas
requests
deep_sort_realtime
inference-sdk
